changed itemValue from float to int

+++ src/Item/ArmorItem.java
@@ -25,7 +25,7 @@ public class ArmorItem extends Item{
 	 * @param itemValue
 	 * @param defenceRating
 	 */
-	public ArmorItem(int id, int x, int y, int width, int height, String name, boolean isVisible, float itemValue, int defenceRating){	
+	public ArmorItem(int id, int x, int y, int width, int height, String name, boolean isVisible, int itemValue, int defenceRating){	
 		super(id, x, y, width, height,name, isVisible, itemValue);
 		this.defenceRating = defenceRating;
 	}
